Better Expansion, Slaughterhouse, and Destroy button
- To make sure the Player will have access to food even if they don't
start close to grain, one of the basic buildings can be a
SlaughterHouse, which will give the town grain once.
- When using the Destroy building button, INSTEAD of destroying lets
disable the old building game object, keep it where it is, and then when
you destroy for example a Metal Wall it will destroy the metal wall and
enable the basic wall it had.
- We will need to keep track of the boosts done by buildings to subtract
them when that building is disabled or destroyed,  then bring them back
if it is enabled