New Towers, Animations by Calcimatrice
New Towers completed give the player 2 options for damage. A fast
machine gun with low damage and fast rate of fire and a cannon with high
damage (AoE) and slow rate of fire.
NEXT: Food system. Some buildings will require food to keep them
working. They require Citizens to be stationed in them, and your
citizens MUST be fed! It's the fist and foremost duty of any proper
Shore Village Captain. If the food requirements are not met the
buildings suffer penalties.