Individual Canvas, Storage Fix, Sorting Layers, and Particles
- Placed a canvas on each building and units to display status
indicators. Want to add indicator text and images to buildings to know
what they are doing (ex. Shooting, Debuffing, Seeking target, Starved,
etc.)
- Units display damage when hit
- Fixed Storage again so it withdraws full ammount of ore/water stored
and hands it to the total. All charges are done to the total, not the
storage. You can still lose whatever is still inside storage.
- Units and Buildings layers are now sorted according to their positions
-Explosion and Hit particles for damage fx. Will probably replace these
with good pixel art animations.