Combat fixes, Pixel Snap material
- tested a few scripts with the Camera, still have to decide what to
keep.
- Fixed combat issue that didn't allow units to be pooled as soon as
they hit 0 hp
- Added new Unit: KAMIKAZE BAT
It flies to a random location and blows up if it collides with anything
in the way. It's like playing battleship with the player's buildings,
guessing where a building will be.
- Adapted Wave Spawner Script to allow for mixing units within a wave
while being able to group them. Currently working on the Spawn
indicators to allow Player to see what type of unit and how many are
incoming.